Added stuff, lots of it. and started on end arena.
Björn there is an explosive barrel for you to work with. I will find or draw an explosion animation and animate it for you in class. I also tweaked some enemy parameters and added them a bit around the map to get the feeling going. Also there is a small problem on their dead state whenever the player returns to a room that has been disabled(they stand up again).